pyproject.toml
[tool.poetry]
name = "book-management-ms"
version = "0.1.0"
description = "book management service"
authors = ["DV"]
package-mode = false
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
psycopg2-binary = "^2.9.10"
alembic = "^1.13.3"
pydantic-settings = "^2.6.0"
uvicorn = "^0.32.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.16"
bcrypt = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
